rename resource folder to feature
