+++ functions-nest/src/app.module.ts
@@ -3,10 +3,10 @@ import { Logger, Module } from '@nestjs/common'
 
 import { FirebaseModule } from './firebase'
 import { AppController } from './app.controller'
-import { LoadoutModule } from './resources/loadout'
-import { ResourceModule } from './resources/resource'
+import { LoadoutModule } from './features/loadout'
+import { ResourceModule } from './features/resource'
 import { Attachment, Clothing, Gear, Weapon } from './entities'
-import { ArmoryModule } from './resources/armory/armory.module'
+import { ArmoryModule } from './features/armory/armory.module'
 import { AuthModule } from './auth/auth.module'
 
 @Module({
