+++ functions-nest/src/features/armory/armory.module.ts
