+++ functions-nest/src/features/armory/armory.service.ts
