+++ functions-nest/src/features/loadout/loadout-clothing.controller.ts
