+++ functions-nest/src/features/loadout/loadout-clothing.service.ts
