+++ functions-nest/src/features/loadout/loadout-gear.controller.ts
