+++ functions-nest/src/features/loadout/loadout-gear.service.ts
