+++ functions-nest/src/features/loadout/loadout-weapon-attachment.controller.ts
