+++ functions-nest/src/features/loadout/loadout-weapon-attachment.service.ts
