+++ functions-nest/src/features/loadout/loadout-weapon.controller.ts
