+++ functions-nest/src/features/loadout/loadout-weapon.service.ts
