+++ functions-nest/src/features/loadout/loadout.service.ts
