+++ functions-nest/src/features/resource/index.ts
