+++ functions-nest/src/features/resource/resource.controller.ts
