+++ functions-nest/src/features/resource/resource.dto.ts
