+++ functions-nest/src/features/resource/resource.module.ts
