+++ functions-nest/src/features/resource/resource.service.ts
